NMR-3359: jcamp parser initial implementation.
This isn't supposed to convert all parameters yet, only to get the parser structure right and convert the binary data properly.
No integration to spinit for the moment, can only be run from JCampToDataset in tests. The goal is to limit dependencies to spinlab
in order to reuse this for NMRfx as well later on.

Squashed from many commits to avoid storing 40Mb of test data and keeping all WIP commits...